+++ CHANGELOG.md
@@ -7,11 +7,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-03-28
+
 ### Added
+
+- Filter by status
+
+## [0.4.0] - 2022-03-28
+
+### Added
+
+- Search by assembly options name
+
+## [0.3.0] - 2022-03-28
+
+### Added
+
+- Pagination on list view
+
+## [0.2.0] - 2022-03-28
+
+### Added
+
+- Assembly Option list view
+
+## [0.1.0] - 2022-03-14
+
+### Added
+
 - Assembly Option graphQL validation added to the Front-End code
 - Base Error and Success Messages defined
 
 ### Added
+
 - Base Assembly Options page and form created
 - Base Key and SKU sections added to the Assembly Form
 - All Intl messages added